day 38: sub class

+++ src/day38_MethodOverriding/Toyota.java
@@ -1,7 +1,7 @@
 package day38_MethodOverriding;
 
 public class Toyota extends Car{
-    public Toyota(String brand, String model, int year, double price, String color, double miles) {
+    public Toyota(String brand, String model, int year, double price, String color, int miles) {
         super(brand, model, year, price, color, miles);
     }
 }
